PR-B-4: extract OverloadResolver (binder-side wrapper around OverloadResolution)
Moves the call-site overload-resolution machinery out of Binder.cs into a
new internal sealed class OverloadResolver at
src/Core/CodeAnalysis/Binding/OverloadResolver.cs, continuing the Phase-1
Binder decomposition (predecessors: PR-B-1 BinderContext, PR-B-2
MemberLookup, PR-B-3 ConversionClassifier). This is a pure mechanical
extraction — no behaviour change, no public-surface change, zero IL
hash diffs on RefactoringBaselineTests.

Methods moved out of Binder.cs into OverloadResolver:
- BindCallExpression(CallExpressionSyntax)
- BindConstructorCallExpression(CallExpressionSyntax, StructSymbol)
- BindExplicitConstructorCallExpression(...)
- BindExtensionFunctionCall(...)
- BindUserInstanceCall(...)
- ExpandParamsArguments(...) and its helper ConvertParamsElement(...)
- BuildOrderedCallArguments (static)
- TryReorderUserCallArguments — both overloads
- TryReorderUserCallArgumentsWithDefaults(...)
- TryAnalyzeCallArgumentLayout(...)
- UnwrapNamedArgumentValue / FirstNamedArgumentName (static helpers)
- TryReportUnknownNamedArgumentForClr(Constructor)(...)
- ValidateRefArguments(...)
- SelectInstanceOverloadOrReport / SelectBestInstanceOverload /
  SelectBestUserOverload / SelectBestUserOverloadCore
- IsApplicableUserCallable (static)
- CreateOptionalUserDefaultArgument (static)
- CreatePossiblyElidedCall
- TryBuildReceiverSubstitution (static)

OverloadResolver wraps the pure reflection-level resolver in
OverloadResolution.cs (which is unchanged). It depends only on
BinderContext (for the diagnostics bag + scope), MemberLookup (for
candidate enumeration), and ConversionClassifier (for per-argument
conversion). It never back-references Binder; the still-on-Binder
helpers it needs are injected as narrow Func / custom-delegate
callbacks in the constructor — same pattern PR-B-3 established.

Deliberately stayed on Binder (deferred to later PRs):
- BindAccessorCall and the rest of the BindAccessor* family — moves
  with B-9 ExpressionBinder
- BindIndirectCallExpression / delegate-invocation binding — B-9
- The CLR-call binding paths (BindClrStaticCall, BindClrInstanceCall,
  TryBindClrConstructorCall, TryBindIntrinsicCall,
  TryBindInheritedClrInstanceCall) — these stay on Binder and are
  invoked by OverloadResolver via callbacks
- The pure static utility helpers used by BindCallExpression
  (InferTypeArguments, SubstituteType, SatisfiesConstraint,
  DescribeConstraint, IsAsyncIteratorReturnType, TryGetFunctionLiteral,
  WrapAsTask, CreateErasedFunctionLiteralAdapter) — 80+ call sites
  outside the moved scope, so they stay on Binder and are passed as
  callbacks (mechanical move-only at this stage)

Untouched: OverloadResolution.cs (the pure resolver), Conversion.cs,
all GS#### diagnostic codes, all public/internal symbols visible to
Compilation.cs, HoverComputer.cs, or any test.

Validation:
- dotnet build GSharp.sln: 0 warnings, 0 errors
- Core.Tests: 2120 passed / 1 skipped (unchanged from baseline)
- RefactoringBaselineTests: 92 passed / 1 skipped — zero IL hash diffs
- LanguageServer.Tests: 242 passed
- Interpreter.Tests: 72 passed
- Compiler.Tests pass cleanly when run with sufficient file-handle
  ulimit; local OOM/file-descriptor exhaustion is environmental
  (16 GB shared Mac), CI will validate the full suite

LoC delta: Binder.cs -2,432 (17,122 → 14,690).
OverloadResolver.cs new: 2,729 LoC.

Next PR in the sequence: B-5 PatternBinder.
